This fixes a race condition in the way SET_PARAMS dispatch is handled.
The problem:
There's a race condition in SET_PARAMS dispatch handling, which
manifests whenever SET_PARAMS is called on different route keys in quick
succession, e.g. if this.props.navigation.setParams() is called during
componentWillMount() of each screen in a TabNavigator

This happens because currently SET_PARAMS dispatch calls setState({ nav
}) for each call. When a lot of SET_PARAMS calls are received in quick
succession, because of the async nature of setState(), the effect is
that of only some of the setState() actually happening. For a test with
4 tabs, only the first and last saw the corresponding state change done.

The solution:
This patch modifies the dispatch handling of SET_PARAMS to ensure that
setState() uses a serialised merge approach to ensure that all the
setState calls see the corresponding changes in the corresponding route keys.